Refactored handling of standard C headers and implementations
Moved functions around to appropriate files. Got rid of subfolders
that were not needed.
