+++ share/smack/include/math.h
